+++ src/app/components/blockchain/Connected/Disconnect.tsx
@@ -1,0 +1,16 @@
+import { useDisconnect } from "wagmi";
+
+const Disconnect = () => {
+	const { disconnect } = useDisconnect();
+	const onClickDisconnect = () => {
+    disconnect();
+  };
+
+	return (
+		<div className="card">
+			<button onClick={onClickDisconnect}>Disconnect</button>
+		</div>
+	)
+}
+
+export default Disconnect;
