+++ src/app/components/blockchain/Connected/index.tsx
@@ -1,22 +1,17 @@
 import { Text } from "@0xsequence/design-system";
-import { useAccount, useDisconnect } from "wagmi";
+import { useAccount } from "wagmi";
 import TestSendTransaction from "./TestSendTransaction";
+import Disconnect from "./Disconnect";
 import ChainInfo from "./ChainInfo";
 
 const Connected = () => {
   const { address, chain } = useAccount();
-  const { disconnect } = useDisconnect();
-  const onClickDisconnect = () => {
-    disconnect();
-  };
   return (
     <>
       <Text variant="large" fontWeight="bold" color="text100">
         Connected with address: {address}
       </Text>
-      <div className="card">
-        <button onClick={onClickDisconnect}>Disconnect</button>
-      </div>
+      <Disconnect />
       {chain && <ChainInfo chain={chain} />}
       <TestSendTransaction />
     </>
